+++ org/w3c/css/selectors/SelectorsList.java
@@ -7,6 +7,11 @@ package org.w3c.css.selectors;
 import org.w3c.css.selectors.attributes.AttributeAny;
 import org.w3c.css.selectors.attributes.AttributeBegin;
 import org.w3c.css.selectors.attributes.AttributeExact;
+import org.w3c.css.selectors.combinators.ChildCombinator;
+import org.w3c.css.selectors.combinators.ColumnCombinator;
+import org.w3c.css.selectors.combinators.DescendantCombinator;
+import org.w3c.css.selectors.combinators.NextSiblingCombinator;
+import org.w3c.css.selectors.combinators.SubsequentSiblingCombinator;
 import org.w3c.css.util.ApplContext;
 import org.w3c.css.util.InvalidParamException;
 import org.w3c.css.util.Messages;
