+++ org/w3c/css/selectors/combinators/ChildCombinator.java
@@ -1,34 +1,36 @@
-//
-// Author: Yves Lafon <ylafon@w3.org>
-//
-// (c) COPYRIGHT MIT, ERCIM, Keio, Beihang, 2021.
+// $Id$
+// Author: Jean-Guilhem Rouel
+// (c) COPYRIGHT MIT, ERCIM and Keio, 2005.
 // Please first read the full copyright statement in file COPYRIGHT.html
-package org.w3c.css.selectors;
+package org.w3c.css.selectors.combinators;
+
+import org.w3c.css.selectors.Selector;
 
 /**
- * Column Combinator
+ * Child<br />
+ * Created: Sep 1, 2005 3:58:00 PM<br />
  */
-public class ColumnCombinator implements Selector {
+public class ChildCombinator implements Selector {
 
     /**
      * @see Selector#toString()
      */
     public String toString() {
-        return " || ";
+        return " > ";
     }
 
     /**
      * @see Selector#getName()
      */
     public String getName() {
-        return "||";
+        return ">";
     }
 
     /**
      * @see Selector#canApply(Selector)
      */
     public boolean canApply(Selector other) {
-        return false;
+        return true;
     }
 
 }
