+++ org/w3c/css/selectors/combinators/ColumnCombinator.java
@@ -1,27 +1,29 @@
-// $Id$
-// Author: Jean-Guilhem Rouel
-// (c) COPYRIGHT MIT, ERCIM and Keio, 2005.
+//
+// Author: Yves Lafon <ylafon@w3.org>
+//
+// (c) COPYRIGHT MIT, ERCIM, Keio, Beihang, 2021.
 // Please first read the full copyright statement in file COPYRIGHT.html
-package org.w3c.css.selectors;
+package org.w3c.css.selectors.combinators;
+
+import org.w3c.css.selectors.Selector;
 
 /**
- * Descendant<br />
- * Created: Sep 1, 2005 3:57:40 PM<br />
+ * Column Combinator
  */
-public class DescendantCombinator implements Selector {
+public class ColumnCombinator implements Selector {
 
     /**
      * @see Selector#toString()
      */
     public String toString() {
-        return " ";
+        return " || ";
     }
 
     /**
      * @see Selector#getName()
      */
     public String getName() {
-        return " ";
+        return "||";
     }
 
     /**
