+++ org/w3c/css/selectors/combinators/DescendantCombinator.java
@@ -2,33 +2,35 @@
 // Author: Jean-Guilhem Rouel
 // (c) COPYRIGHT MIT, ERCIM and Keio, 2005.
 // Please first read the full copyright statement in file COPYRIGHT.html
-package org.w3c.css.selectors;
+package org.w3c.css.selectors.combinators;
+
+import org.w3c.css.selectors.Selector;
 
 /**
- * Child<br />
- * Created: Sep 1, 2005 3:58:00 PM<br />
+ * Descendant<br />
+ * Created: Sep 1, 2005 3:57:40 PM<br />
  */
-public class ChildCombinator implements Selector {
+public class DescendantCombinator implements Selector {
 
     /**
      * @see Selector#toString()
      */
     public String toString() {
-        return " > ";
+        return " ";
     }
 
     /**
      * @see Selector#getName()
      */
     public String getName() {
-        return ">";
+        return " ";
     }
 
     /**
      * @see Selector#canApply(Selector)
      */
     public boolean canApply(Selector other) {
-        return true;
+        return false;
     }
 
 }
