+++ org/w3c/css/selectors/combinators/NextSiblingCombinator.java
@@ -2,7 +2,9 @@
 // Author: Jean-Guilhem Rouel
 // (c) COPYRIGHT MIT, ERCIM and Keio, 2005.
 // Please first read the full copyright statement in file COPYRIGHT.html
-package org.w3c.css.selectors;
+package org.w3c.css.selectors.combinators;
+
+import org.w3c.css.selectors.Selector;
 
 /**
  * Adjacent<br />
