+++ org/w3c/css/selectors/combinators/SubsequentSiblingCombinator.java
@@ -1,7 +1,9 @@
 // $Id$
 // (c) COPYRIGHT MIT, ERCIM and Keio, 2009.
 // Please first read the full copyright statement in file COPYRIGHT.html
-package org.w3c.css.selectors;
+package org.w3c.css.selectors.combinators;
+
+import org.w3c.css.selectors.Selector;
 
 /**
  * GeneralSibling<br />
